fix: let form submit when formNoValidate is used on submitter
This commit ensures that a form submitted with a button using the
attribute `formNoValidate` always get submitted even if the validation
would fail. As Typescript DOM types document:

> `formNoValidate` Overrides any validation or required attributes on a
form or form elements to allow it to be submitted without validation.
This can be used to create a "save draft"-type submit option.